snp data is now loaded and read from mongodb
if the snp data is not fully loaded in the database, the data will be loaded in upon starting the server. this process will take a bit more than a minute, and the website will not be able to run analysis during this period. upon finish, a message will be printed to the server console signalling the end of loading the snp data. then, as long as the database isn't cleared, starting the server should be much faster than before.