Add thres and hover (#191)
* add api routes that load project data to mongo and retrieve aggregated data.

* Alter mongo routes to take result ID

* edit development files

gulpfile.js has a mybuild task which allows for bundling of css
package.json supports needed dependencies
start_mongo.sh stores the mongodatabase in a local dbpath folder

* integrate new matrix visualization with project

Now, the website automatically loads compiled data to mongo and shows the new matrix visualization instead of the old one. (The old matrix visualization is still in the project but no longer has a route)

* small fixes

* Refactor

* refactor for new material-ui version finishes

* fixed mui theme

* load data after analysis run instead of visualization mount

* small edits

* switch from localdb to postgresql

* authentication for postgres

* disabled horizontal scrolling of topaxis and grid in visualization

* ignore db

* gulpfile functional

* 30 row problem and discoloration problem fixed

* row height change and topaxis color change

* snp data is now loaded and read from mongodb

if the snp data is not fully loaded in the database, the data will be loaded in upon starting the server. this process will take a bit more than a minute, and the website will not be able to run analysis during this period. upon finish, a message will be printed to the server console signalling the end of loading the snp data. then, as long as the database isn't cleared, starting the server should be much faster than before.

* a little cleanup

* add emo user

* implement mongo in backend support

* file path is passed to c++ instead of an object

* results now are loaded to mongo directly from c++ backend

* demo user words correctly

before, a demo user was created each time the server restarted. now only one demo user is created

* small fix

* result correlation data and index positions are now both loaded to mongo from the c++ backend.

previously, only result correlation data was loaded to mongo from c++, and index data was loaded later in javascript. now that both are loaded up at the same time, storing result data should be faster

* add Plink

* reformat webapp

* reformat...

* load Plink trait name

* add thresh query parameter support to /api/get-range route

* traits on y-axis label instead of just row number

* trait labels are abbreviated and shown in full when hovered over

* got rid of zoombar text, labels properly cover cells

* imporve the Installation

* edit the Instruction

* delete run_db.sh

* make it only one bash

* visualization backend api returns min and max values

previously, sorting made mongo timeout because it could not handle so much data

* I return tod424a1bd98188793ad6023bedfbcc2cdebe56682 and Then I add the threshold and fake hover.

* remove unnecessary package import

* update scripts

* update scripts and add stop scripts

* update db container names in scripts

* update script messages

* Moving to webpack

* extra webpack integration things

* persist redux store on localstorage to not lose project files

* deleted unnecessary launch_docker_with_dbs.sh

* pressing the canceljob button actually cancels the job in the backend